+++ .gitignore
@@ -38,8 +38,6 @@
 !/app/assets/builds/.keep
 
 /node_modules
-
-# ここから下2行を追記
-.DS_Store
 /vendor/bundle
-# ここまで追記.DS_Store
+
+.DS_Store
