+++ Procfile.dev
@@ -1,3 +1,4 @@
 web: env RUBY_DEBUG_OPEN=true bin/rails server -b 0.0.0.0 -p 3000
 js: yarn build --watch
-css: yarn build:css --watch
+css: yarn watch:css
+css: bin/rails dartsass:watch
