+++ app/assets/config/manifest.js
@@ -1,3 +1,2 @@
 //= link_tree ../images
 //= link_tree ../builds
-//= link simulator.css
