+++ app/assets/stylesheets/application.scss
@@ -1,0 +1,1 @@
+// Sassy
