+++ app/assets/stylesheets/application.tailwind.css
@@ -1,4 +1,3 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
