--- app/helpers/rooms_helper.rb
@@ -1,2 +1,0 @@
-module RoomsHelper
-end
