+++ app/views/bookmark/create.html.erb
@@ -1,0 +1,2 @@
+<h1>Bookmark#create</h1>
+<p>Find me in app/views/bookmark/create.html.erb</p>
