+++ app/views/bookmark/destroy.html.erb
@@ -1,0 +1,2 @@
+<h1>Bookmark#destroy</h1>
+<p>Find me in app/views/bookmark/destroy.html.erb</p>
