+++ app/views/rooms/top.html.erb
@@ -1,7 +1,8 @@
-<div class="top-wrapper">
-  <div class="top-inner-text">
-    <%= link_to "Floor Simulator", floor_simulator_path %>
-    <%= link_to "Wallpaper Simulator", wallpaper_simulator_path %>
-    <%= link_to "sign up", new_user_path %>
+<div class="container mx-auto py-8">
+  <h2 class="text-2xl font-semibold text-gray-700 mb-6">シミュレーションを選択してください</h2>
+  <div class="flex space-x-4">
+    <%= link_to "Floor Simulator", floor_simulator_path, class: "bg-gray-200 hover:bg-gray-300 text-gray-800 py-2 px-4 rounded" %>
+    <%= link_to "Wallpaper Simulator", wallpaper_simulator_path, class: "bg-gray-200 hover:bg-gray-300 text-gray-800 py-2 px-4 rounded" %>
+    <%= link_to "Sign Up", new_user_path, class: "bg-blue-500 hover:bg-blue-600 text-white py-2 px-4 rounded" %>
   </div>
 </div>
