+++ app/views/shared/_before_login_header.html.erb
@@ -1,4 +1,6 @@
-<header>
-  <h1>ログイン前のヘッダー</h1>
-  <%= link_to 'ログイン', login_path %>
+<header class="bg-gray-100 p-4 shadow-md">
+  <h1 class="text-xl font-semibold text-gray-700">Love Your Home</h1>
+  <nav class="mt-2">
+    <%= link_to 'ログイン', login_path, class: "text-base text-gray-600 hover:text-gray-900" %>
+  </nav>
 </header>
