+++ app/views/shared/_footer.html.erb
@@ -1,11 +1,5 @@
-<footer id='footer' class='footer-bottom'>
-  <div class='container'>
-    <div class='row'>
-      <div class='col-sm-6 col-12'>
-        <div class='copyright'>
-          <p>Copyright © 2024 hase_n</p>
-        </div>
-      </div>
-    </div>
+<footer class="bg-gray-100 text-gray-600 py-4 mt-8 border-t border-gray-300">
+  <div class="container mx-auto text-center">
+    <p>&copy; 2024 hase_n</p>
   </div>
 </footer>
