+++ app/views/user_sessions/new.html.erb
@@ -1,22 +1,23 @@
-<div class="container">
-  <div class="row">
-    <div class=" col-md-10 col-lg-8 mx-auto">
-      <h1>ログイン</h1>
-      <%= form_with url: login_path do |f| %>
-        <div class="mb-3">
-          <%= f.label :email %>
-          <%= f.email_field :email, class: "form-control" %>
-        </div>
-        <div class="mb-3">
-          <%= f.label :password %>
-          <%= f.password_field :password, class: "form-control" %>
-        </div>
-        <%= f.submit "ログイン", class: "btn btn-primary" %>
-      <% end %>
-      <div class='text-center'>
-        <%= link_to '登録ページへ', new_user_path %>
-        <%= link_to 'パスワードをお忘れの方はこちら', '#' %>
-      </div>
+<div class="max-w-lg mx-auto mt-10 p-6 bg-white shadow-md rounded-md">
+  <h1 class="text-2xl font-semibold text-gray-800 text-center mb-6">ログイン</h1>
+  <%= form_with url: login_path, class: "space-y-6" do |f| %>
+    <div>
+      <%= f.label :email, "メールアドレス", class: "block text-sm font-medium text-gray-700" %>
+      <%= f.email_field :email, class: "mt-1 block w-full px-4 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring focus:ring-beige-500 focus:border-beige-500" %>
     </div>
+    
+    <div>
+      <%= f.label :password, "パスワード", class: "block text-sm font-medium text-gray-700" %>
+      <%= f.password_field :password, class: "mt-1 block w-full px-4 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring focus:ring-beige-500 focus:border-beige-500" %>
+    </div>
+
+    <div class="text-center">
+      <%= f.submit "ログイン", class: "w-full bg-beige-400 hover:bg-beige-500 text-white font-bold py-2 px-4 rounded-md transition duration-200" %>
+    </div>
+  <% end %>
+
+  <div class="mt-4 text-center space-x-4">
+    <%= link_to '登録ページへ', new_user_path, class: "text-sm text-gray-600 hover:underline" %>
+    <%= link_to 'パスワードをお忘れの方はこちら', '#', class: "text-sm text-gray-600 hover:underline" %>
   </div>
 </div>
