+++ config/application.rb
@@ -1,7 +1,6 @@
 require_relative "boot"
 
 require "rails/all"
-require 'sassc-rails'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -24,11 +23,5 @@ module Myapp
     #
     # config.time_zone = "Central Time (US & Canada)"
     # config.eager_load_paths << Rails.root.join("extras")
-    # 追記↓
-    config.generators do |g|
-      g.skip_routes true
-      g.helper false
-      g.test_framework nil
-    end
   end
 end
