+++ config/environments/production.rb
@@ -99,7 +99,4 @@ Rails.application.configure do
   # ]
   # Skip DNS rebinding protection for the default health check endpoint.
   # config.host_authorization = { exclude: ->(request) { request.path == "/up" } }
-
-  # このプロセスでは、scssファイル内のSass特有の構文や機能（変数、ネスト、ミックスインなど）が標準のCSSコードに変換される
-  config.assets.css_compressor = nil
 end
