+++ config/initializers/dartsass.rb
@@ -1,0 +1,3 @@
+Rails.application.config.dartsass.builds = {
+  "rooms.scss"  => "rooms.css"
+}
