+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.2].define(version: 2024_09_01_142643) do
+ActiveRecord::Schema[7.2].define(version: 2024_09_09_191430) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
