+++ tailwind.config.js
@@ -4,5 +4,13 @@ module.exports = {
     './app/helpers/**/*.rb',
     './app/assets/stylesheets/**/*.css',
     './app/javascript/**/*.js'
-  ]
+  ],
+  theme: {
+    extend: {
+      colors: {
+        beige: '#f5f5dc',  // ソフトなベージュ色
+      },
+    },
+  },
+  plugins: [],
 }
