--- test/controllers/users_controller_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class UsersControllerTest < ActionDispatch::IntegrationTest
-  # test "the truth" do
-  #   assert true
-  # end
-end
